# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Veylora Global API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /couriers/{consNo}/track-updates:
    get:
      operationId: listTrackUpdates
      tags: [track-updates]
      summary: List tracking history for a shipment
      parameters:
        - name: consNo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tracking updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackUpdate"
    post:
      operationId: addTrackUpdate
      tags: [track-updates]
      summary: Add a tracking status update
      parameters:
        - name: consNo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackUpdateInput"
      responses:
        "201":
          description: Tracking update added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackUpdate"
  /track-updates/{id}:
    delete:
      operationId: deleteTrackUpdate
      tags: [track-updates]
      summary: Delete a tracking update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # ─── Officers (Admin) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResult:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── Auth ───────────────────────────────────────────────────────────────
...
    TrackUpdate:
      type: object
      required:
        - id
        - cid
        - consNo
        - updateDate
        - currentCity
        - newStatus
        - comments
        - currentLocation
        - bkTime
      properties:
        id:
          type: integer
        cid:
          type: integer
        consNo:
          type: string
        updateDate:
          type: string
        currentCity:
          type: string
        newStatus:
          type: string
        comments:
          type: string
        currentLocation:
          type: string
        bkTime:
          type: string
        rName:
          type: ["string", "null"]
        rPhone:
          type: ["string", "null"]
        rMail:
          type: ["string", "null"]
        rAdd:
          type: ["string", "null"]
    TrackUpdateInput:
      type: object
      required:
        - updateDate
        - currentCity
        - newStatus
        - comments
        - currentLocation
        - bkTime
      properties:
        updateDate:
          type: string
        currentCity:
          type: string
        newStatus:
          type: string
        comments:
          type: string
        currentLocation:
          type: string
        bkTime:
          type: string
        rName:
          type: string
        rPhone:
          type: string
        rMail:
          type: string
        rAdd:
          type: string

    # ─── Officer ─────────────────────────────────────────────────────────────
...