# lib/api-zod/src/generated/types/trackUpdateInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Veylora Global API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackUpdateInput {
  updateDate: string;
  currentCity: string;
  newStatus: string;
  comments: string;
  currentLocation: string;
  bkTime: string;
  rName?: string;
  rPhone?: string;
  rMail?: string;
  rAdd?: string;
}